# benchmark + opt: make Folding::apply slightly faster