--- a/src/common/folding.hh
+++ b/src/common/folding.hh
@@ -28,6 +28,7 @@ enum {
 /// Applies the folding algorithm to each character in the given string,
 /// making another one as a result.
 wstring apply( wstring const &, bool preserveWildcards = false );
+wstring apply2( wstring const &, bool preserveWildcards = false );
 
 /// Applies only simple case folding algorithm. Since many dictionaries have
 /// different case style, we interpret words differing only by case as synonyms.