@shenlebantongying 2024-11-12
Is there any reason why we shouldn't use Qt's case folding, they use the same data file but newer https://github.com/qt/qtbase/blob/dev/util/unicode/data/CaseFolding.txt